requirements.txt
alembic==1.4.1
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
azure-ai-formrecognizer==3.2.0b2
azure-common==1.1.28
azure-core==1.22.1
azure-storage-blob==12.11.0
backoff==1.11.1
boto3==1.22.1
botocore==1.25.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
cloudpickle==2.0.0
cryptography==3.4.6
databricks-cli==0.16.6
dill==0.3.4
dnspython==2.2.1
docker==5.0.3
docopt==0.6.2
elastic-apm==6.9.1
elasticsearch==7.10.0
entrypoints==0.4
farm-haystack==1.3.0
fastapi==0.75.2
filelock==3.6.0
Flask==2.1.1
gitdb==4.0.9
GitPython==3.1.27
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
huggingface-hub==0.5.1
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
inflect==5.5.2
isodate==0.6.1
itsdangerous==2.1.2
jarowinkler==1.0.2
Jinja2==3.1.1
jmespath==1.0.0
joblib==1.1.0
jsonschema==4.4.0
langdetect==1.0.9
loguru==0.6.0
lxml==4.8.0
Mako==1.2.0
MarkupSafe==2.1.1
mlflow==1.13.1
mmh3==3.0.0
monotonic==1.6
more-itertools==8.12.0
msrest==0.6.21
networkx==2.8
nltk==3.7
num2words==0.5.10
numpy==1.21.6
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pinecone-client==2.0.9
posthog==1.4.7
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.1
protobuf==3.20.1
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pyparsing==3.0.8
pyrsistent==0.18.1
pysrt==1.1.2
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.20.0
python-editor==1.0.4
-e git+https://github.com/ms-mousa/pytube-working-v12.git@04d4ff5a3e5055ac7a97b5395c7e285ff3ac654d#egg=pytube
pytz==2022.1
PyYAML==6.0
quantulum3==0.7.10
querystring-parser==1.2.4
rapidfuzz==2.0.11
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.8.0
semantic-version==2.9.0
sentence-transformers==2.2.0
sentencepiece==0.1.96
seqeval==1.2.2
setuptools-rust==1.3.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
SQLAlchemy==1.4.36
SQLAlchemy-Utils==0.38.2
sqlparse==0.4.2
starlette==0.17.1
tabulate==0.8.9
threadpoolctl==3.1.0
tika==1.24
tokenizers==0.10.3
torch==1.10.2
torchvision==0.12.0
tqdm==4.64.0
transformers==4.13.0
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websocket-client==1.3.2
websockets==10.3
Werkzeug==2.1.1
zipp==3.8.0
